functions/requirements.txt
firebase_functions~=0.1.0
pillow==10.3.0
requests==2.32.2
tweepy==4.14.0
openai==0.28
python-dotenv==1.0.1
requests==2.32.2